Split atlassian connectors
